feat: add Day 2 project - Tip Calculator
This commit adds the main.py file for Day 2 of the tutorial series. The script is a simple tip calculator that asks for the bill amount, the tip % and the number of people paying the bill. It then gives how much each person will have to pay on that bill including tip.